requirements.txt
lxml==3.1.0
cssselect==0.7.1
dumptruck==0.1.0
scraperwiki-local==0.1.0
